Added css and html update
Added comments to both CSS and HTML files. The html rule is for the whole page. The rule is for the list items that in the unordered list that is in the nav and has a class of horizontal. The display is block.  The rule is for the first list item and it has a margin to the left of 2. The rule is for the anchor tags that are n the unordered list and in the id tag rightcol. When the inks are hovered by the cursor, the text turns black. Ready for review.